add sequence generator for service, money and userRole

+++ src/main/java/ua/com/foxminded/accountingsystem/web/controller/admin/ServiceController.java
@@ -1,4 +1,4 @@
-package ua.com.foxminded.accountingsystem.web.controller;
+package ua.com.foxminded.accountingsystem.web.controller.admin;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ua.com.foxminded.accountingsystem.model.Currency;
-import ua.com.foxminded.accountingsystem.model.Price;
+import ua.com.foxminded.accountingsystem.model.Money;
 import ua.com.foxminded.accountingsystem.model.Service;
 import ua.com.foxminded.accountingsystem.service.ServiceService;
 
@@ -34,7 +34,7 @@ public class ServiceController {
 
     @GetMapping
     public String getAllServices(Model model){
-        List<Service> services = serviceService.getAllServices();
+        List<Service> services = serviceService.findAllServices();
         log.debug("Found: "+services.size()+" services!");
         model.addAttribute("services", services);
         return "admin/services";
@@ -42,13 +42,16 @@ public class ServiceController {
 
     @GetMapping("/remove/{id}")
     public String remove(@PathVariable Long id){
-        serviceService.remove(id);
+        log.debug("Remove service with id: "+id);
+        if (id != null) {
+            serviceService.remove(id);
+        }
         return "redirect:/admin/services";
     }
 
     @GetMapping("/{id}")
     public String getOneService(@PathVariable Long id, Model model) {
-        model.addAttribute("service",serviceService.getById(id));
+        model.addAttribute("service",serviceService.findById(id));
         model.addAttribute("currencies", Currency.values());
         return "admin/service";
     }
@@ -61,18 +64,18 @@ public class ServiceController {
 
     @GetMapping("/newService")
     public String newService(Model model){
-        List<Price> prices = new ArrayList<>();
+        List<Money> monies = new ArrayList<>();
         for (Currency currency : Arrays.asList(Currency.values())) {
-            Price price = new Price();
-            price.setCurrency(currency);
-            prices.add(price);
+            Money money = new Money();
+            money.setCurrency(currency);
+            monies.add(money);
         }
 
-        Price employeeRate = new Price();
+        Money employeeRate = new Money();
         employeeRate.setCurrency(Currency.UAH);
 
         Service service = new Service();
-        service.setPrices(prices);
+        service.setMonies(monies);
         service.setEmployeeRate(employeeRate);
 
         model.addAttribute("service", service);
